Made R function platform independent.

+++ R/assignTiebreaks.r
@@ -20,21 +20,44 @@ for(i in 1:nrow(tiebreaks))
     }
 }
 
-if(sum(is.na(saveList$out.assign))>0)
+if (.Platform$OS.type=="unix")
 {
-	naout <- subset(names(saveList$out.assign),is.na(saveList$out.assign))
-	write.table(as.numeric(naout),file=(paste(pupil.dir,"NAs.csv",sep="")),row.names=FALSE,col.names=FALSE)
-	
-	if(file.exists(paste(pupil.dir,"unassigned_for_cluster",sep=""))){unlink(paste(pupil.dir,"unassigned_for_cluster",sep=""),recursive=TRUE)}
-	dir.create(paste(pupil.dir,"unassigned_for_cluster",sep=""))
-	
-	data <- read.csv(paste(pupil.dir,"acoustic_data.csv",sep=""),header=TRUE)
-	
-	for(name in subset(names(saveList$out.assign),is.na(saveList$out.assign)))
+	if(sum(is.na(saveList$out.assign))>0)
 	{
-		name.assign <- paste("%0",nchar(max(as.numeric(rownames(data)))),"s",sep="")
-		name.out <- sprintf(name.assign,name)
-		file.copy(from=paste(pupil.dir,"cut_syllables/",name.out,".wav",sep=""),to=paste(pupil.dir,"unassigned_for_cluster/",name.out,".wav",sep=""))
+		naout <- subset(names(saveList$out.assign),is.na(saveList$out.assign))
+		write.table(as.numeric(naout),file=(paste(pupil.dir,"NAs.csv",sep="")),row.names=FALSE,col.names=FALSE)
+	
+		if(file.exists(paste(pupil.dir,"unassigned_for_cluster",sep=""))){unlink(paste(pupil.dir,"unassigned_for_cluster",sep=""),recursive=TRUE)}
+		dir.create(paste(pupil.dir,"unassigned_for_cluster",sep=""))
+	
+		data <- read.csv(paste(pupil.dir,"acoustic_data.csv",sep=""),header=TRUE)
+	
+		for(name in subset(names(saveList$out.assign),is.na(saveList$out.assign)))
+		{
+			name.assign <- paste("%0",nchar(max(as.numeric(rownames(data)))),"s",sep="")
+			name.out <- sprintf(name.assign,name)
+			file.copy(from=paste(pupil.dir,"cut_syllables/",name.out,".wav",sep=""),to=paste(pupil.dir,"unassigned_for_cluster/",name.out,".wav",sep=""))
+		}
+	}
+}else if (.Platform$OS.type=="windows"){
+	
+	if(sum(is.na(saveList$out.assign))>1)
+	{
+		naout <- subset(names(saveList$out.assign),is.na(saveList$out.assign))
+		write.table(as.numeric(naout),file=(paste(pupil.dir,"NAs.csv",sep="")),row.names=FALSE,col.names=FALSE)
+	
+		if(file.exists(paste(pupil.dir,"unassigned_for_cluster",sep=""))){unlink(paste(pupil.dir,"unassigned_for_cluster/",sep=""),recursive=TRUE)}
+		dir.create(paste(pupil.dir,"unassigned_for_cluster",sep=""))
+	
+		data <- read.csv(paste(pupil.dir,"acoustic_data.csv",sep=""),header=TRUE)
+	
+		for(name in subset(names(saveList$out.assign),is.na(saveList$out.assign)))
+		{
+			name.assign <- paste("%0",nchar(max(as.numeric(rownames(data)))),"s",sep="")
+			name.out <- sprintf(name.assign,name)
+	        name.out <- gsub(" ","0",name.out)
+			file.copy(from=paste(pupil.dir,"cut_syllables/",name.out,".wav",sep=""),to=paste(pupil.dir,"unassigned_for_cluster/",name.out,".wav",sep=""))
+		}
 	}
 }
